requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.11.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
phonenumbers==8.13.38
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
